+++ src/index.d.ts
@@ -30,6 +30,8 @@ export type StoreCreator<S, A> = (reducer: Reducer<S, A>, initialState: S) => St
 
 export type Enhancer<S, A> = (creator: StoreCreator<S, A>) => StoreCreator<S, A>;
 
+type AnyEnhancer = unknown;
+
 export type CreateGlobalState = <S extends {}, A extends {}>(initialState: S) => {
   GlobalStateProvider: React.ComponentType<GlobalStateProviderProps>;
   useGlobalState: UseGlobalState<S>;
@@ -39,20 +41,8 @@ export type CreateGlobalState = <S extends {}, A extends {}>(initialState: S) =>
 export type CreateStore = <S extends {}, A extends {}>(
   reducer: Reducer<S, A>,
   initialState: S,
-  enhancer?: Enhancer<S, A>,
+  enhancer?: Enhancer<S, A> | AnyEnhancer,
 ) => Store<S, A>;
 
 export const createGlobalState: CreateGlobalState;
 export const createStore: CreateStore;
-
-// for patch redux
-
-export type MiddlewareAPI<S, A> = {
-  getState: () => S;
-  dispatch: Dispatch<A>;
-};
-
-export type Middleware<S, A> =
-  (store: MiddlewareAPI<S, A>) => (next: Dispatch<A>) => (action: A) => A;
-
-export type ApplyMiddleware<S, A> = (...args: Middleware<S, A>[]) => Enhancer<S, A>;